application/requirements.txt
mysql-connector
scapy
prometheus_client
